# app/Models/Subscription.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Model;

class Subscription extends Model
{
    //
    protected $fillable =
    [
        'name',
        'tel',
        'email',
    ];
}