# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API softskills",
        "description": "Documentation API softskills",
        "contact": {
            "name": "Kouame Banh",
            "email": "kouamebanh@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8001/api/v2",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v2/subscription/list": {
            "get": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Get list of subscriptions",
                "description": "Returns list of subscriptions",
                "operationId": "getSubscriptions",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Subscription"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subscription not found"
                    },
                    "500": {
                        "description": "Error retrieving subscriptions"
                    }
                }
            }
        },
        "/api/v2/subscription": {
            "post": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Add a new subscription",
                "description": "Add a new subscription",
                "operationId": "addSubscription",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Subscription"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Subscription"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error creating subscription"
                    }
                }
            }
        },
        "/api/v2/update/subscription/{id}": {
            "put": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Update an existing subscription",
                "description": "Update an existing subscription",
                "operationId": "updateSubscription",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subscription ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/Subscription"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscription updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Subscription"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subscription not found"
                    },
                    "500": {
                        "description": "Error updating subscription"
                    }
                }
            }
        },
        "/api/v2/delete/subscription/{id}": {
            "delete": {
                "tags": [
                    "Subscriptions"
                ],
                "summary": "Delete an existing subscription",
                "description": "Delete an existing subscription",
                "operationId": "deleteSubscription",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Subscription ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Subscription deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Subscription"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Subscription not found"
                    },
                    "500": {
                        "description": "Error deleting subscription"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Subscription": {
                "required": [
                    "name",
                    "email",
                    "tel"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": " B4TtD@example.com"
                    },
                    "tel": {
                        "type": "string",
                        "example": "1234567890"
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Utilisez un jeton JWT pour l'authentification.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}